Orders complete

+++ database/migrations/2024_12_20_092958_create_trips_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->timestamp('departure_scheduled_at');
             $table->string('destination', 255);
             $table->timestamp('arrival_scheduled_at');
-            $table->unsignedInteger('price');
+            $table->decimal('price', 4, 2);
             $table->timestamps();
         });
     }
